// src/models/collection-detail.model.ts
export interface CollectionDetail {
    oneDayVolume: number;
    name: string;
    image: string;
    mintDate: number;
    mintDateDiffFromNow: number;
    floorPrice: number;
    listedCount: number;
    volume: number;
    supply: number;
    lastSold: number;
    uniqueHolders: number;
    magicEdenFloorPrice: number;
    magicEdenVolume: number;
    magicEdenListedCount: number;
}